refactor: rename files for consistency with atoms table
