+++ libs/frontend/domain/type/src/use-cases/types/get-types/useTypesTableData.ts
@@ -1,7 +1,7 @@
 import type { ITypeService } from '@codelab/frontend/abstract/core'
 import { useAsyncFn } from 'react-use'
 
-export const useGetTypesTableData = (typeService: ITypeService) => {
+export const useTypesTableData = (typeService: ITypeService) => {
   const [{ loading: isLoadingAllTypes }, getAllTypes] = useAsyncFn(
     typeService.getAll.bind(typeService),
     [],
